# docs/handsUPApiSpec.yml
openapi: 3.0.3
info:
  title: Hands Up - OpenAPI 3.0
  description: |-
   This is a simple service contract for HandsUp API.
  contact:
    email: tmkdt.cos301@gmail.com
  version: 1.0.0
servers:
  - url: http://localhost:2000/handsup #example url
...
paths:
...
  /learning/progress/{username}:
    get:
      tags:
        - learning
      summary: Get learning progress
      description: Retrieve stats about the user's learning progress
      operationId: learningProgress
      parameters:
        - name: username
          in: path
          required: true
          description: The username of the user to retrieve progress for
          schema:
            type: string
      responses:
        '200':
          $ref: '#/components/responses/successStatsResponse'
        '400':
          $ref: '#/components/responses/notFoundResponse'
        '404':
          $ref: '#/components/responses/badRequestResponse'
        '500':
          $ref: '#/components/responses/internalServerError'
...
components:
  responses:
    internalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            type: object
            properties:
              status:
                type: string
                example: "error"
              message:
                type: string
                example: "Internal server error"
    successStatsResponse:
      description: Operation was successful
      content:
        application/json:
          schema:
            type: object
            properties:
              status:
                type: string
                example: "success"
              message:
                type: string
                example: "Retrieval completed successfully"
              data:
                type: array
                items:
                  type: object
                  properties:
                    lessonsCompleted:
                      type: integer
                      example: 10
                    difficultyLevel:
                      type: integer
                      example: 2
...
    notFoundResponse:
      description: User not found
      content:
        application/json:
          schema:
            type: object
            properties:
              status:
                type: string
                example: "error"
              message:
                type: string
                example: "Resource not found"
...
    badRequestResponse:
      description: Bad request
      content:
        application/json:
          schema:
            type: object
            properties:
              status:
                type: string
                example: "error"
              message:
                type: string
                example: "Bad request"